recv_request_head(): ensure data is terminated
Previously, the function might have found a CRLFCRLF sequence in the
uninitialized part of buffer. It assumed, that buffer is initialized to
zeros, when it was not.

This fix makes it terminate the read data after each subsequent read
from the socket, so that strstr() doesn't look at uninitialized memory.